restructured the code a bit
